+++ msm8226/libgralloc/gpu.h
@@ -35,7 +35,7 @@ class gpu_context_t : public alloc_device_t {
     gpu_context_t(const private_module_t* module,
                   IAllocController* alloc_ctrl);
 
-    int gralloc_alloc_buffer(size_t size, int usage,
+    int gralloc_alloc_buffer(unsigned int size, int usage,
                              buffer_handle_t* pHandle,
                              int bufferType, int format,
                              int width, int height);
@@ -44,7 +44,7 @@ class gpu_context_t : public alloc_device_t {
 
     int alloc_impl(int w, int h, int format, int usage,
                    buffer_handle_t* pHandle, int* pStride,
-                   size_t bufferSize = 0);
+                   unsigned int bufferSize = 0);
 
     static int gralloc_alloc(alloc_device_t* dev, int w, int h,
                              int format, int usage,
