+++ msm8226/libgralloc/gr.h
@@ -33,7 +33,7 @@
 struct private_module_t;
 struct private_handle_t;
 
-inline size_t roundUpToPageSize(size_t x) {
+inline unsigned int roundUpToPageSize(unsigned int x) {
     return (x + (PAGE_SIZE-1)) & ~(PAGE_SIZE-1);
 }
 
@@ -47,16 +47,16 @@ inline Type ALIGN(Type x, Type align) {
 
 int mapFrameBufferLocked(struct private_module_t* module);
 int terminateBuffer(gralloc_module_t const* module, private_handle_t* hnd);
-size_t getBufferSizeAndDimensions(int width, int height, int format, int usage,
-                                  int& alignedw, int &alignedh);
-size_t getBufferSizeAndDimensions(int width, int height, int format,
-                                  int& alignedw, int &alignedh);
+unsigned int getBufferSizeAndDimensions(int width, int height, int format,
+        int usage, int& alignedw, int &alignedh);
+unsigned int getBufferSizeAndDimensions(int width, int height, int format,
+        int& alignedw, int &alignedh);
 
 
 // Attributes include aligned width, aligned height, tileEnabled and size of the buffer
 void getBufferAttributes(int width, int height, int format, int usage,
                            int& alignedw, int &alignedh,
-                           int& tileEnabled, size_t &size);
+                           int& tileEnabled, unsigned int &size);
 
 
 bool isMacroTileEnabled(int format, int usage);
